move styles out of components dir
